chore: move redirect dataset links to api module
