+++ bd_api/apps/core/urls.py
@@ -2,8 +2,6 @@
 from django.urls import include, path, re_path
 from django.views.generic import RedirectView, TemplateView
 
-from bd_api.apps.core.views import DatasetRedirectView
-
 
 def render_robots():
     return TemplateView.as_view(template_name="robots.txt", content_type="text/plain")
@@ -13,6 +11,4 @@ urlpatterns = [
     path("robots.txt", render_robots()),
     re_path(r"^healthcheck/", include("health_check.urls")),
     path("", RedirectView.as_view(url="admin", permanent=True), name="home"),
-    #
-    path("dataset_redirect/", DatasetRedirectView.as_view(), name="dataset_redirect"),
 ]
